refactor(iOS, Paper): do not add subviews to header config in HostTree (#2553)
## Description

I was just writing #2552 and got confused that the header config
subviews are added as subviews to the header config view
in host tree, before they are attached to the navigation bar view
hierarchy. This transient state does not make any sense,
and we do not do similar thing on Fabric.

## Changes

`self.superview` returns now `nil` until the subview is mounted in
navigation bar view hierarchy.

Also moved the function to Paper specific section of file, because it is
not called on Fabric on any codepath.

## Test code and steps to reproduce

We haven't used this behaviour. There should be no regression in text
examples.

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes