fix: take preload into account when freezing children (#2566)
Fixed https://github.com/software-mansion/react-native-screens/issues/2557. We
have to take preloaded screens into account when freezing screens. Till
now, the assumption was that the newest screens are the ones rendered,
right now top children can be the preloaded ones, so we have to filter
them from calculating the size of frozen screens. We could also think of
freezing the preloaded screens, wdyt about it @satya164 ?