Handle cases when subview is added after initial render
I moved the state update logic from screen stack to config & subviews.
I've also made HeaderConfig implement `RCTMountingTransactionObserving` protocol
where it now requests layout if necessary.

I've also discovered a bug where after removal of one of header subviews
the layout is not reset to the same one as it would be when rendered w/o
the removed subview in the first place.