fix: Make `NativeViewHierarchyManager` nullable (#2643)
The `NativeViewHierarchyManager` argument in `.addUIBlock` is nullable.
(See
[`UIBlock::execute(...)`](https://github.com/facebook/react-native/blob/864c34e9f6c55319417633206e58cefc3c962e7d/packages/react-native/ReactAndroid/src/main/java/com/facebook/react/uimanager/UIBlock.kt#L12))

This PR fixes that definition and makes it nullable, and also checks for
null. In our case, we likely want to exit out of that and just do
nothing. (but lmk, I can also change to a throw)

<!--
Description and motivation for this PR.

Include Fixes #<number> if this is fixing some issue.

Fixes # .
-->

- Fixed `NativeViewHierarchyManager` being nullable

<!--
Please describe things you've changed here, make a **high level**
overview, if change is simple you can omit this section.

For example:

- Updated `about.md` docs

-->

<!--

Here you can add screenshots / GIFs documenting your change.

You can add before / after section if you're changing some behavior.

-->

<!--
Please include code that can be used to test this change and short
description how this example should work.
This snippet should be as minimal as possible and ready to be pasted
into editor (don't exclude exports or remove "not important" parts of
reproduction example)
-->

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [x] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacper.kafara@swmansion.com>