fix(iOS): fix quick modal presentation sequence where foreign modal is being dismissed (#2671)
## Description

Closes #2668 

See the #2668 for issue description.

Making long story short: The issue involves cases when we have two
modals `A, b`, where `A` is owned and `b` is foreign, and we
dismiss `b` and present another owned modal `B` in single transaction or
in separate transactions but pulled quickly enough so that previous
transitions (caused by previous transaction) are not finished yet.

Currently, because of `b` being dismissed by its host view controller,
we would wrongfully dismiss the controller underneath (`A`). Now the
code takes possibility of `b` being in dismissal into account.

## Changes

We now check whether the potentially foreign modal is being dismissed -
if so, we schedule our updates to run **after** dismissal transition
completes.

## Test code and steps to reproduce

Tested for regressions: 

* [x] `Modal` screen in Example
[PR](https://github.com/software-mansion/react-native-screens/pull/1996)
* [x] Test2048
[PR](https://github.com/software-mansion/react-native-screens/pull/2113)
* [x] Test1829
[PR](https://github.com/software-mansion/react-native-screens/pull/1912)
* [x] Test1299
[PR](https://github.com/software-mansion/react-native-screens/pull/1326)

This issue is quite similar to #1299 but concerns foreign modal not
owned one and mechanism are a bit different, therefore I've added
`Test2668`.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes