fix(Android,Fabric): prevent another infinite state update loop for header subviews with zero size (#2696)
## Description

Closes #2675

### Error mechanism

This one was tricky to figure out.

Few concurrent facts first:

1. When `statusBarTranslucent` is `true` we apply `paddingTop` to native
`Toolbar` to heighten the header.
2. Since 4.6.0 we update frames of header subviews in ShadowTree to the
values from HostTree, everytime the frame changes in HostTree.
3. We had a check in header subview shadow node that would prevent node
size update if the frame was (0, 0).

Header subview that represents the search bar has no content. Therefore
native layout sets it frame to exactly (0, 0), but with origin `(x,
toolbar.paddingTop)`.
This frame - `((x, toolbar.paddingTop), (0, 0))` is send to shadow node,
**where it is ignored**, but the layout is triggered by subsequent
*header config* update.
Therefore Yoga resolves height of the subview to full height of the
parent (usually `154 px = 40 dip`). New layout metrics are sent to
HostTree, where next native toolbar layout
determines toolbar size to be its content height + padding == subview
height + paddingTop. This gets send to ShadowTree, then back to
HostTree, native layout is triggered and another
`paddingTop` value is added to the overall header height, and so on. 

This is a regression introduced in 4.6.0 and limited to Fabric. 

The fix is simple - accept the (0, 0) size for the subview in
ShadowTree. We just need to use more reasonable value to denote the
uninitialized frame - `{-1, -1}` seems like a better choice as
it is an invalid frame.

## Changes

- **Use different initial value in state so that we can unambiguously
distinguish between un- and initialized state**

## Test code and steps to reproduce

`Test2675`

Tested on all combinations:

* Android SDK 34 (w/o edge-to-edge) and 35 (w/ edge-to-edge enabled),
* `statusBarTranslucent: true` and `false`,
* search bar present, not present
* other header elements present / not present.

Also tested iOS on the same example, because the code changes affect
shared C++ code.


## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes