feat: support for RN 0.78.0-rc.5 (#2626)
## Description

Add library support for react-native 0.78 & use 0.78 in example apps.
 
> [!caution]
This PR breaks compatibility with older react native versions on new
architecture. In case it is merged new architecture will be supported
only from 0.78 onwards. The breaking change in core we need to follow:
>
> * https://github.com/facebook/react-native/pull/48062
> 
> There are also breaking changes made in core regarding method /
function signatures, see required adjustments here
https://github.com/software-mansion/react-native-screens/pull/2626/commits/5a7e4407ea761cd9761032286797330a4d5d5106
>
> **I believe these Android method signature changes do break backward
compatibility also on old architecture!**


## Changes


## Test code and steps to reproduce

<!--
Please include code that can be used to test this change and short
description how this example should work.
This snippet should be as minimal as possible and ready to be pasted
into editor (don't exclude exports or remove "not important" parts of
reproduction example)
-->

## Checklist

- [ ] Ensured that CI passes

---------

Co-authored-by: Marc Rousavy <me@mrousavy.com>