fix(Android): missing content when using flex for formsheet contents (#2748)
## Description

Fixing the styles applied to the content in screens with `formSheet`
presentation.
Currently when using `flex: 1` on the content of the sheet it might
cause the content to disappear. This PR mitigates that.


## Changes

When not in `fitToContents` mode we apply `flex: 1` to content wrapper. 


## Test code and steps to reproduce

`TestFormSheet` - notice that the contents has `flex: 1` and it displays
correctly.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes