fix(Android): prevent crash by properly managing lifecycle of the dimming view (#2750)
## Description

When we have a `formSheet` on Android & we navigate to regular push
screen on top of it & we go back `navigation.goBack()`
the app will crash.
This is because the `ScreenStackFragment`' view will be recreated, but
it will reuse existing dimming view, leading to a crash when
inserting this dimming view to a coordinator, because no one never
cleared previous parent of it.


## Changes

See the code, basically I've ensured that the previous dimming view is
invalidated & on fragment view recreation (when it is reattached)
new dimming view is created.


## Test code and steps to reproduce

`TestFormSheet`

1. Open sheet
2. Navigate to second
3. Go back
4. In this very moment app crashes


## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes