fix(Android): findTitleTextViewInToolbar couldn't find the correct view in certain cases (#2757)
## Description

Fixes #2745 

## Changes

Use `TextUtils` to compare the value of `view.text` and `toolbar.title`.
`text` and `title` are `CharSequence`, there are no guarantee that they
will be the same type, therefore the `==` comparison could fail.

For example, `Toolbar` could be styled natively in the project,
therefore the value of its `TextView` child could be `SpannableString`,
while title could be still just a `String`.

## Test code and steps to reproduce

## Checklist

- [ ] Ensured that CI passes

Co-authored-by: Herli Halim <herli.halim@zip.co>