fix(Android,native-stack): fix check for `sheetAllowedDetents` value & apply appropriate styles (#2762)
## Description

I've made a small change in v5 code in #2748, however due to my overlook
it introduced a regression in native-stack v5.
Since it is a quick fix I'm patching it. 

Basically I've parsed & transformed `sheetAllowedDetents` value in
`RouteView`, **before** passing it to `MaybeNestedStack`, where
I've checked for `sheetAllowedDetents === 'fitToContents'`, which could
not be true, because in such case it has been transformed to `[-1]`
literal in `RouteView`.

Only the latest 4.10.0-beta.0 & beta.1 versions are affected.

## Changes

Removed `sheetAllowedDetents` prop value transformation in `RouteView`
since `Screen` component handles it nicely internally.

## Test code and steps to reproduce

`TestFormSheet`, but you need to run it using
`createNativeStackNavigator` from `react-native-screens/native-stack`.
Don't forget to change the `presentation` prop to `stackPresentation`,
when changing to v5.

## Checklist

- [ ] Ensured that CI passes