fix: update workaround preventing React Native from hiding frozen screens to work with RN 0.78 (#2778)
## Description

React Native sets `display: 'none'` for frozen screens - this behavior
can become an issue, if you for example press back button twice very
fast on iOS - frozen screen might still be invisible. The problem and
its workaround suggested by @grahammendick can be seen in issue
https://github.com/software-mansion/react-native-screens/issues/1207.
The workaround was introduced in PR
https://github.com/software-mansion/react-native-screens/pull/1208.

Recently, React Native 0.78 has changed the way of diffing props and the
workaround stopped working. The details and a new workaround has been
suggested by @grahammendick once again in issue
https://github.com/software-mansion/react-native-screens/issues/2753 and
PR https://github.com/grahammendick/navigation/pull/860.

This PR introduces the suggested workaround.

A big thank you to @grahammendick for figuring it out and suggesting the
solution once again, appreciate it!

Furthermore, the new workaround fixes a bug with back button on the
screen stack header on iOS. The bug was probably introduced with the
update to 0.78. The back button would disappear and then reappear after
a short delay with generic "Back" label instead of the screen name.


https://github.com/user-attachments/assets/a371df20-d230-4e38-8770-022b72bacb6a


https://github.com/user-attachments/assets/8c3c9572-f77b-46ab-804c-d50361eeba06

Fixes
https://github.com/software-mansion/react-native-screens/issues/2753.

## Changes

- change `display: false` to `display: null`,
- change `ViewConfig` interface to allow `null` value.

## Tests

Run `TestFreeze` example, click `Go to Details` twice and then click the
back button in the header. The button behaves as expected, which was not
the case before - the button would disappear and then reappear with a
generic "Back" title.

## Checklist

- [x] Updated TS types
- [ ] Ensured that CI passes