chore: remove unnecessary enableFreeze from Example.tsx (#2779)
## Description

In `Example.tsx`, we run `enableFreeze()` but it will be overwritten by
`enableFreeze(true)` in `App.tsx` as the call from `App.tsx` is run
after `Example.tsx`.

In case changing this setting is necessary for testing, it should be
changed in top-level file of the app, in our case it is the `App.tsx`.

I left out `enableFreeze` wrapped in `useEffect` in `Test2229.tsx` as it
will be called when running the test.

## Changes

- remove `enableFreeze()` from `Example.tsx`

## Checklist

- [ ] Ensured that CI passes