Move configuration to `onCreateView` & sheet configuration to SheetDelegate
The sheet must be configured in `onCreateView`, because
`onAttachedToWindow` is called before `onViewCreated` for
reattached fragments and apparently some behaviour code runs,
because later configuration is not reflected on screen immediately.

I also need the sheet delegate to be created & initialized before
the behaviour is configured. Therefore I'm splitting creation &
configuration into two phases and also moving configuration code
to the delegate to unclutter `ScreenStackFragment`.`