fix(Android): `formSheet` with `fitToContents` does not have correct height after goBack (#2789)
## Description

When fragment is reattached (on navigation back) `onCreateView` is
called which tries to configure the form sheet (that has `fitToContents`
set).
On view recreation the subtree is not laid out (the flag is cleared) &
therefore current code fails.

## Changes

We now utilise the fact, that the Screen's subtree is retained by
react-native and the components are not recreated themselves -> old
height is stored in the member variable.
We reuse it. 


## Test code and steps to reproduce

Added Test2789

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes