fix(iOS): use prefersPageSizing for modal on iOS 18.0+ (#2797)
## Description

Brings back `stackPresentation: 'modal'`'s size on iOS 18.0+.

Since the release of iOS 18, screens with `stackPresentation: 'modal'`
changed size on devices with a bigger screen, such as an iPad - they
became much smaller, which impacted applications of the library's users
(see issue
https://github.com/software-mansion/react-native-screens/issues/2549).
This happens because `UIModalPresentationAutomatic`, which is used by
`react-native-screens` for modal on iOS, has been changed with the
release of iOS 18. `UIModalPresentationAutomatic` is now mapped to
`UIModalPresentationFormSheet` for iOS >=18 and
`UIModalPresentationPageSheet` for earlier versions (this was the
default before).

For more context please see PR
https://github.com/software-mansion/react-native-screens/pull/2793 but
the TL;DR is that we consider this a breaking change for our users and
therefore we decided to bring back the old behavior for
`react-native-screens` v4 using
[`prefersPageSizing`](https://developer.apple.com/documentation/uikit/uisheetpresentationcontroller/preferspagesizing?language=objc)
property of the `UISheetPresentationController` which has been available
since iOS 17.0+.

Resolves
https://github.com/software-mansion/react-native-screens/issues/2549.

## Changes

- use `prefersPageSizing` for modal on iOS 18.0+

## Screenshots / GIFs

### Before
![Modal without `prefersPageSizing` on iOS
18](https://github.com/user-attachments/assets/724bd109-b0c0-47e4-b897-43cd53ef3dd9)

### After
![Modal with `prefersPageSizing` on iOS
18](https://github.com/user-attachments/assets/b0925934-25dc-4103-8422-8fb07a18fd18)

## Test code and steps to reproduce

Open `Modals` example screen and use `Open modal` button.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacper.kafara@swmansion.com>