chore: update iOS min supported version to 15.1 (#2799)
## Description

This is not a breaking change, because we support `react-native` 0.76+
anyway, and it has iOS min
deployment target [set to
15.1](https://github.com/react-native-community/discussions-and-proposals/discussions/812)

This is needed because I work with @maciejstosio on an issue with
`backButtonDisplayMode` & it has been introduced
in iOS 14.0. The workaround I've found works only on iOS >= 14.0,
therefore it would be nice to not have to
add any ifdefs in source code. 

## Changes

Bumped the min. supported iOS version to 15.1 from 11.0.

I've left the new-arch-condition in the podspec code, despite the fact
that it is not needed,
so that it is clear where do we change things when it becomes needed.

tvOS version is left unchanged for now (most likely it should also be
updated!).

## Test code and steps to reproduce

The app should build and run on any device with 15.1 or newer iOS
version.

## Checklist

- [ ] Ensured that CI passes