feat(iOS): add pageSheet presentation for native-stack (#2793)
## Description

Adds new presentation for Native Stack - `pageSheet`
(`UIModalPresentationPageSheet` modal style on iOS, `modal` on Android).

Since the release of iOS 18, screens with `stackPresentation: 'modal'`
changed size on devices with a bigger screen, such as an iPad - they
became much smaller, which impacted applications of the library's users
(see issue
https://github.com/software-mansion/react-native-screens/issues/2549).
This happens because `UIModalPresentationAutomatic`, which is used by
`react-native-screens` for `modal` on iOS, has been changed with the
release of iOS 18. `UIModalPresentationAutomatic` is now mapped to
`UIModalPresentationFormSheet` for iOS >=18 and
`UIModalPresentationPageSheet` for earlier versions (this was the
default before).

~~Apple added SwiftUI API
[(link)](https://developer.apple.com/documentation/swiftui/presentationsizing)
to allow to choose the behavior of the modal (`form`, `page`, `fitted`
or a custom one) but I wasn't able to find an equivalent functionality
in UIKit.~~ It turns out there is a new property in
`UISheetPresentationController`, see
[here](https://github.com/software-mansion/react-native-screens/pull/2793#pullrequestreview-2705276787)

We considered 3 possible solutions to this problem:
- changing modal presentation style on iOS from
`UIModalPresentationAutomatic` to `UIModalPresentationPageSheet` for
`stackPresentation: 'modal'` to bring back old behavior for all modals:
- some users might've already designed their apps with the smaller sheet
in mind so changing the style might break their apps,
- adding a feature flag that would allow users to choose whether screens
with `stackPresentation: 'modal'` should use
`UIModalPresentationAutomatic` or `UIModalPresentationPageSheet`
- this solution would provide the easiest fix for the users impacted by
the change,
- even though 2 different sheet sizes would be available, users would
need to choose only one of them for their application,
- adding a new `stackPresentation` that would use
`UIModalPresentationPageSheet` on iOS and a regular `modal` on Android:
- users can choose whether they want to use `Automatic` or `PageSheet`
behavior per-screen basis,
- hopefully, fixing old applications with some find-and-replace magic
won't be difficult.

We decided on the third option.

tvOS does not allow `UIModalPresentationPageSheet` so it fallbacks to
`UIModalPresentationFullScreen` (the same behavior as
`UIModalPresentationAutomatic`).

This change will require changes to `react-navigation/native-stack` as
well.

Resolves
https://github.com/software-mansion/react-native-screens/issues/2549.

## Changes

- add support for the new presentation in native and JS code
- add `pageSheet` examples in `Stack Presentation` and `Modals` example
screens

## Screenshots / GIFs

### `modal`

![modal](https://github.com/user-attachments/assets/657d0130-dfa0-4433-8f11-d8eb32a6146b)

### `pageSheet`

![pageSheet](https://github.com/user-attachments/assets/4759b662-119a-4886-82c4-6712502d3b6a)

## Test code and steps to reproduce

Open `Stack Presentation`, `Modals` example screens and open `pageSheet`
(at the moment, you might need to manually add `pageSheet` in
react-navigation files `packages/native-stack/src`,
`packages/native-stack/src/utils/getModalRoutesKeys.ts`,
`packages/native-stack/src/views`).

## Checklist

- [x] Included code example that can be used to test this change
- [x] Updated TS types
- [x] Updated documentation: <!-- For adding new props to native-stack
-->
- [x]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [x]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [x]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [x]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes