test(e2e,Fabric): add e2e tests for issue/PR examples 42..528 (#2787)
## Description

Check which example screens from issues/PRs can be used in e2e testing
for tests `Test42, ..., Test528` and implement them if possible for
Fabric. In the future, each non-skipped test will be in a seperate PR.

### Test42
Skipped because we can’t really check what is the orientation of the
screen with Detox.

### Test111
Skipped because we can’t test flickering with Detox.

### Test263
Skipped because example differs from the problem mentioned in
https://github.com/software-mansion/react-native-screens/issues/263.
Even when I recreated the issue, it still occurs. [This comment from the
PR](https://github.com/software-mansion/react-native-screens/issues/263#issuecomment-930127450)
explains why it is the case.

### Test349
Skipped because we can’t test autofill with Detox. The test also uses
outdated prop name.

### Test364
Skipped because the test uses `tabBarVisible` prop that no longer
exists. `react-navigation` documentation suggests other
[solution](https://reactnavigation.org/docs/hiding-tabbar-in-screens/)
to this problem.

### Test 432
Test created. It checks whether correct squares are fully visible. For
some reason, on Android, these squares are not `100%` visible from the
start (but I don’t see the problem in manual testing). For the test,
`waitingFor` the squares to become `100%` visible works.

### Test 528
Test created but only for iOS because Detox supports changing
orientation of the device [only on
iOS](https://wix.github.io/Detox/docs/api/device/#devicesetorientationorientation).
Test checks if the headerRight button is visible after changing
screens/orientation.

## Changes

- add `Test432` and `Test528`
- add comments for every test screen in `apps/src/tests/index.ts` with
the reason for (not) implementing e2e test for it

## Test code and steps to reproduce

CI

## Checklist

- [x] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>